fix(entities): use bigint for monetary/quantity columns, extend OrderItem with SyncableEntity

+++ backend/src/inventory/entities/inventory-adjustment.entity.ts
@@ -11,7 +11,7 @@ export class InventoryAdjustment {
   @Column('uuid')
   productId: string;
 
-  @Column('integer')
+  @Column('bigint')
   quantityInGrams: number;
 
   @Column({ default: '' })
